+++ package.json
@@ -1,5 +1,5 @@
 {
-  "version": "3.2.0",
+  "version": "3.3.0",
   "name": "@bloomprotocol/qr-react",
   "description": "SVG QR Renderer For React",
   "author": "Bloom Team <team@bloom.co>",
