+++ src/core/modes/QRNum.ts
@@ -1,49 +1,49 @@
 import { QRData } from './QRData'
-import { Mode } from '../enums'
+import { Mode } from '../shared'
 import { BitBuffer } from '../util/BitBuffer'
+
+const chatToNum = (character: string) => {
+  if (character >= '0' && character <= '9') {
+    return character.charCodeAt(0) - '0'.charCodeAt(0)
+  }
+
+  throw new Error(`Unsupported character: ${character}`)
+}
+
+const strToNum = (str: string) => {
+  let num = 0
+
+  for (let i = 0; i < str.length; i += 1) {
+    num = num * 10 + chatToNum(str.charAt(i))
+  }
+
+  return num
+}
 
 export class QRNum extends QRData {
   constructor(data: string) {
     super(Mode.NUM, data)
   }
 
-  public write(buffer: BitBuffer) {
+  write(buffer: BitBuffer) {
     const { data } = this
     let i = 0
 
     while (i + 2 < data.length) {
-      buffer.put(QRNum.strToNum(data.substring(i, i + 3)), 10)
+      buffer.put(strToNum(data.substring(i, i + 3)), 10)
       i += 3
     }
 
     if (i < data.length) {
       if (data.length - i === 1) {
-        buffer.put(QRNum.strToNum(data.substring(i, i + 1)), 4)
+        buffer.put(strToNum(data.substring(i, i + 1)), 4)
       } else if (data.length - i === 2) {
-        buffer.put(QRNum.strToNum(data.substring(i, i + 2)), 7)
+        buffer.put(strToNum(data.substring(i, i + 2)), 7)
       }
     }
   }
 
-  public getLength() {
+  getLength() {
     return this.data.length
-  }
-
-  private static strToNum(str: string) {
-    let num = 0
-
-    for (let i = 0; i < str.length; i += 1) {
-      num = num * 10 + QRNum.chatToNum(str.charAt(i))
-    }
-
-    return num
-  }
-
-  private static chatToNum(character: string) {
-    if (character >= '0' && character <= '9') {
-      return character.charCodeAt(0) - '0'.charCodeAt(0)
-    }
-
-    throw new Error(`Unsupported character: ${character}`)
   }
 }
