+++ src/core/util/Polynomial.ts
@@ -3,7 +3,7 @@ import { glog, gexp } from './math'
 export class Polynomial {
   private num: number[]
 
-  public constructor(num: number[], shift = 0) {
+  constructor(num: number[], shift = 0) {
     let offset = 0
 
     while (offset < num.length && num[offset] === 0) {
@@ -20,33 +20,11 @@ export class Polynomial {
     return this.num.length
   }
 
-  public getAt(index: number) {
+  getAt(index: number) {
     return this.num[index]
   }
 
-  public toString() {
-    let buffer = ''
-    for (let i = 0; i < this.length; i += 1) {
-      if (i > 0) {
-        buffer += ','
-      }
-      buffer += this.getAt(i)
-    }
-    return buffer.toString()
-  }
-
-  public toLogString() {
-    let buffer = ''
-    for (let i = 0; i < this.length; i += 1) {
-      if (i > 0) {
-        buffer += ','
-      }
-      buffer += glog(this.getAt(i))
-    }
-    return buffer.toString()
-  }
-
-  public multiply(e: Polynomial) {
+  multiply(e: Polynomial) {
     const num = new Array<number>(this.length + e.length - 1)
 
     for (let i = 0; i < this.length; i += 1) {
@@ -59,7 +37,7 @@ export class Polynomial {
     return new Polynomial(num)
   }
 
-  public mod(e: Polynomial): Polynomial {
+  mod(e: Polynomial): Polynomial {
     if (this.length - e.length < 0) {
       return this
     }
